projects/01-travel-assistant/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-assistant"
version = "0.1.0"
description = "A multi-step task travel assistant (learning project)."
requires-python = ">=3.9"
dependencies = [
  "requests",
  "tavily-python",
  "openai",
  "urllib3<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

